Added error-report.html, Reconstructed errorhandling in dal classes
- the html page error-report only displays logresults of error related
logs.
- Added attributes to the pdo object from parent class Dal. The
attributes configure the pdo object so it throws an exception with any
type of pdo error.
- All child dal classes replaced there if else exception handling with a
try catch. With the combination of the extra attributes givn to the pdo
object, it can now catch all errors and handle them properly.
- The getResult() method in the Feedback class now checks whether it's
returning null or a value, in the case of null it returns the string
'n/a'.